remove deepmind and rename to custom.gemmascope
